pyproject.toml
[project]
name = "test-framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "camelot-py[base]>=1.0.0",
    "ghostscript>=0.7",
    "openai>=1.72.0",
    "pandas>=2.2.3",
    "pydantic[email,timezone]>=2.11.2",
]
